Fix image creation and mount on 64 bit: changed index type to fixed uint32
This fixes the image creation on systems where long is 64 bit, in such cases the index couldn't be loaded with the fuse implementation.
Not sure if this is the way it was supposed, fixing the fuse implementation (with its index format case distinction) would be an other option, but 32 bit for the index arra (containing only block lengths) should be enough.